@@ -3,7 +3,7 @@
 import logging
 import os
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple, Type, TypeVar, Union
+from typing import Any, Optional, TypeVar, Union
 
 from emmet.core.structure import MoleculeMetadata
 from monty.dev import requires
@@ -21,13 +21,12 @@ try:
 except ImportError:
     cclib = None
 
-__all__ = ["TaskDocument"]
 
 logger = logging.getLogger(__name__)
 _T = TypeVar("_T", bound="TaskDocument")
 
 
-class TaskDocument(MoleculeMetadata):
+class TaskDocument(MoleculeMetadata, extra="allow"):  # type: ignore[call-arg]
     """
     Definition of a cclib-generated task document.
 
@@ -41,37 +40,35 @@ class TaskDocument(MoleculeMetadata):
     logfile: str = Field(
         None, description="Path to the log file used in the post-processing analysis"
     )
-    attributes: Dict = Field(
+    attributes: dict = Field(
         None, description="Computed properties and calculation outputs"
     )
-    metadata: Dict = Field(
+    metadata: dict = Field(
         None,
         description="Calculation metadata, including input parameters and runtime "
         "statistics",
     )
     task_label: str = Field(None, description="A description of the task")
-    tags: List[str] = Field(None, description="Optional tags for this task document")
+    tags: list[str] = Field(None, description="Optional tags for this task document")
     last_updated: str = Field(
         default_factory=datetime_str,
         description="Timestamp for this task document was last updated",
     )
-    _schema: str = Field(
-        __version__,
-        description="Version of atomate2 used to create the document",
-        alias="schema",
+    schema: str = Field(
+        __version__, description="Version of atomate2 used to create the document"
     )
 
     @classmethod
     @requires(cclib, "The cclib TaskDocument requires cclib to be installed.")
     def from_logfile(
-        cls: Type[_T],
+        cls: type[_T],
         dir_name: Union[str, Path],
-        logfile_extensions: Union[str, List[str]],
+        logfile_extensions: Union[str, list[str]],
         store_trajectory: bool = False,
-        additional_fields: Optional[Dict[str, Any]] = None,
-        analysis: Optional[Union[str, List[str]]] = None,
+        additional_fields: Optional[dict[str, Any]] = None,
+        analysis: Optional[Union[str, list[str]]] = None,
         proatom_dir: Optional[Union[Path, str]] = None,
-    ) -> _T:
+    ) -> "TaskDocument":
         """
         Create a TaskDocument from a log file.
 
@@ -123,7 +120,7 @@ class TaskDocument(MoleculeMetadata):
 
         logger.info(f"Getting task doc from {logfile}")
 
-        additional_fields = {} if additional_fields is None else additional_fields
+        additional_fields = additional_fields or {}
 
         # Let's parse the log file with cclib
         cclib_obj = ccread(logfile, logging.ERROR)
@@ -248,7 +245,7 @@ class TaskDocument(MoleculeMetadata):
             metadata=metadata,
         )
         doc.molecule = final_molecule
-        return doc.copy(update=additional_fields)
+        return doc.model_copy(update=additional_fields)
 
 
 @requires(cclib, "cclib_calculate requires cclib to be installed.")
@@ -257,7 +254,7 @@ def cclib_calculate(
     method: str,
     cube_file: Union[Path, str],
     proatom_dir: Union[Path, str],
-) -> Optional[Dict[str, Any]]:
+) -> Optional[dict[str, Any]]:
     """
     Run a cclib population analysis.
 
@@ -354,8 +351,8 @@ def cclib_calculate(
 
 
 def _get_homos_lumos(
-    moenergies: List[List[float]], homo_indices: List[int]
-) -> Tuple[List[float], Optional[List[float]], Optional[List[float]]]:
+    moenergies: list[list[float]], homo_indices: list[int]
+) -> tuple[list[float], Optional[list[float]], Optional[list[float]]]:
     """
     Calculate the HOMO, LUMO, and HOMO-LUMO gap energies in eV.
 
