@@ -1,19 +1,24 @@
 """Job to prerelax a structure using an MD Potential."""
 
-from typing import List, Optional
+from typing import Optional
 
+from ase.stress import voigt_6_to_full_3x3_stress
+from ase.units import GPa
+from emmet.core.math import Matrix3D, Vector3D
 from emmet.core.structure import StructureMetadata
 from pydantic import BaseModel, Extra, Field
 from pymatgen.core.structure import Structure
 from pymatgen.io.ase import AseAtomsAdaptor
 
 
-class IonicStep(BaseModel, extra=Extra.allow):  # type: ignore
+class IonicStep(BaseModel, extra=Extra.allow):  # type: ignore[call-arg]
     """Document defining the information at each ionic step."""
 
     energy: float = Field(None, description="The free energy.")
-    forces: List[List[float]] = Field(None, description="The forces on each atom.")
-    stress: List[float] = Field(None, description="The stress on the lattice.")
+    forces: Optional[list[list[float]]] = Field(
+        None, description="The forces on each atom."
+    )
+    stress: Optional[Matrix3D] = Field(None, description="The stress on the lattice.")
     structure: Structure = Field(None, description="The structure at this step.")
 
 
@@ -26,16 +31,13 @@ class InputDoc(BaseModel):
         description="Whether cell lattice was allowed to change during relaxation.",
     )
     steps: int = Field(
-        None,
-        description="Maximum number of steps allowed during relaxation.",
+        None, description="Maximum number of steps allowed during relaxation."
     )
-    relax_kwargs: dict = Field(
-        None,
-        description="Keyword arguments that passed to the relaxer function.",
+    relax_kwargs: Optional[dict] = Field(
+        None, description="Keyword arguments that passed to the relaxer function."
     )
-    optimizer_kwargs: dict = Field(
-        None,
-        description="Keyword arguments passed to the relaxer's optimizer.",
+    optimizer_kwargs: Optional[dict] = Field(
+        None, description="Keyword arguments passed to the relaxer's optimizer."
     )
 
 
@@ -51,18 +53,18 @@ class OutputDoc(BaseModel):
         description="Energy per atom of the final structure in units of eV/atom.",
     )
 
-    forces: List[List[float]] = Field(
+    forces: Optional[list[Vector3D]] = Field(
         None,
         description="The force on each atom in units of eV/A for the final structure.",
     )
 
     # NOTE: units for stresses were converted to kbar (* -10 from standard output)
     #       to comply with MP convention
-    stress: List[float] = Field(
+    stress: Optional[Matrix3D] = Field(
         None, description="The stress on the cell in units of kbar (in Voigt notation)."
     )
 
-    ionic_steps: List[IonicStep] = Field(
+    ionic_steps: list[IonicStep] = Field(
         None, description="Step-by-step trajectory of the structural relaxation."
     )
 
@@ -110,14 +112,14 @@ class ForceFieldTaskDocument(StructureMetadata):
         relax_kwargs: dict = None,
         optimizer_kwargs: dict = None,
         ionic_step_data: tuple = ("energy", "forces", "magmoms", "stress", "structure"),
-    ):
+    ) -> "ForceFieldTaskDocument":
         """
         Create a ForceFieldTaskDocument for a Task that has ASE-compatible outputs.
 
         Parameters
         ----------
         forcefield_name : str
-            Name of the forcefield used.
+            Name of the force field used.
         result : dict
             The outputted results from the task.
         relax_cell : bool
@@ -133,10 +135,13 @@ class ForceFieldTaskDocument(StructureMetadata):
         """
         trajectory = result["trajectory"].__dict__
 
-        # NOTE: units for stresses were converted to kbar (* -10 from standard output)
-        # to comply with MP convention
+        # NOTE: units for stresses were converted from ev/Angstrom³ to kbar
+        # (* -1 from standard output)
+        # and to 3x3 matrix to comply with MP convention
         for i in range(len(trajectory["stresses"])):
-            trajectory["stresses"][i] = trajectory["stresses"][i] * -10
+            trajectory["stresses"][i] = voigt_6_to_full_3x3_stress(
+                trajectory["stresses"][i] * -10 / GPa
+            )
 
         species = AseAtomsAdaptor.get_structure(trajectory["atoms"]).species
 
