@@ -1,14 +1,14 @@
 """Settings for atomate2."""
+from __future__ import annotations
 
 import warnings
 from pathlib import Path
-from typing import Literal, Optional, Tuple, Union
+from typing import Literal, Optional, Union
 
-from pydantic import BaseSettings, Field, root_validator
+from pydantic import Field, model_validator
+from pydantic_settings import BaseSettings, SettingsConfigDict
 
 _DEFAULT_CONFIG_FILE_PATH = "~/.atomate2.yaml"
-
-__all__ = ["Atomate2Settings"]
 
 
 class Atomate2Settings(BaseSettings):
@@ -31,7 +31,12 @@ class Atomate2Settings(BaseSettings):
     SYMPREC: float = Field(
         0.1, description="Symmetry precision for spglib symmetry finding."
     )
-    CUSTODIAN_SCRATCH_DIR: str = Field(
+    BANDGAP_TOL: float = Field(
+        1e-4,
+        description="Tolerance for determining if a material is a semiconductor or "
+        "metal",
+    )
+    CUSTODIAN_SCRATCH_DIR: Optional[str] = Field(
         None, description="Path to scratch directory used by custodian."
     )
 
@@ -45,7 +50,9 @@ class Atomate2Settings(BaseSettings):
     VASP_NCL_CMD: str = Field(
         "vasp_ncl", description="Command to run non-collinear version of VASP."
     )
-    VASP_VDW_KERNEL_DIR: str = Field(None, description="Path to VDW VASP kernel.")
+    VASP_VDW_KERNEL_DIR: Optional[str] = Field(
+        None, description="Path to VDW VASP kernel."
+    )
     VASP_INCAR_UPDATES: dict = Field(
         default_factory=dict, description="Updates to apply to VASP INCAR files."
     )
@@ -64,16 +71,16 @@ class Atomate2Settings(BaseSettings):
     VASP_CUSTODIAN_MAX_ERRORS: int = Field(
         5, description="Maximum number of errors to correct before custodian gives up"
     )
-    VASP_STORE_VOLUMETRIC_DATA: Optional[Tuple[str]] = Field(
+    VASP_STORE_VOLUMETRIC_DATA: Optional[tuple[str]] = Field(
         None, description="Store data from these files in database if present"
     )
     VASP_STORE_ADDITIONAL_JSON: bool = Field(
-        True,
+        default=True,
         description="Ingest any additional JSON data present into database when "
         "parsing VASP directories useful for storing duplicate of FW.json",
     )
     VASP_RUN_BADER: bool = Field(
-        False,
+        default=False,
         description="Whether to run the Bader program when parsing VASP calculations."
         "Requires the bader executable to be on the path.",
     )
@@ -83,6 +90,14 @@ class Atomate2Settings(BaseSettings):
         description="Determine if the files in folder are being compressed. If True "
         "all the files are compressed. If 'atomate' only a selection of files related "
         "to the simulation will be compressed. If False no file is compressed.",
+    )
+    VASP_INHERIT_INCAR: bool = Field(
+        default=False,
+        description="Whether to inherit INCAR settings from previous calculation. "
+        "This might be useful to port Custodian fixes to child jobs but can also be "
+        "dangerous e.g. when switching from GGA to meta-GGA or relax to static jobs."
+        "Can be overridden on a per-job basis via the inherit_incar keyword of "
+        "VaspInputGenerator.",
     )
 
     LOBSTER_CMD: str = Field(
@@ -104,7 +119,7 @@ class Atomate2Settings(BaseSettings):
         "cp2k.psmp", description="Command to run the MPI version of cp2k"
     )
     CP2K_RUN_BADER: bool = Field(
-        False,
+        default=False,
         description="Whether to run the Bader program when parsing CP2K calculations."
         "Requires the bader executable to be on the path.",
     )
@@ -131,17 +146,17 @@ class Atomate2Settings(BaseSettings):
     CP2K_CUSTODIAN_MAX_ERRORS: int = Field(
         5, description="Maximum number of errors to correct before custodian gives up"
     )
-    CP2K_STORE_VOLUMETRIC_DATA: Optional[Tuple[str]] = Field(
+    CP2K_STORE_VOLUMETRIC_DATA: Optional[tuple[str]] = Field(
         None, description="Store data from these files in database if present"
     )
     CP2K_STORE_ADDITIONAL_JSON: bool = Field(
-        True,
+        default=True,
         description="Ingest any additional JSON data present into database when "
         "parsing CP2K directories useful for storing duplicate of FW.json",
     )
 
     CP2K_ZIP_FILES: Union[bool, Literal["atomate"]] = Field(
-        True,
+        default=True,
         description="Determine if the files in folder are being compressed. If True "
         "all the files are compressed. If 'atomate' only a selection of files related "
         "to the simulation will be compressed. If False no file is compressed.",
@@ -153,17 +168,15 @@ class Atomate2Settings(BaseSettings):
     )
 
     # AMSET settings
-    AMSET_SETTINGS_UPDATE: dict = Field(
+    AMSET_SETTINGS_UPDATE: Optional[dict] = Field(
         None, description="Additional settings applied to AMSET settings file."
     )
 
-    class Config:
-        """Pydantic config settings."""
+    model_config = SettingsConfigDict(env_prefix="atomate2_")
 
-        env_prefix = "atomate2_"
-
-    @root_validator(pre=True)
-    def load_default_settings(cls, values):
+    @model_validator(mode="before")
+    @classmethod
+    def load_default_settings(cls, values) -> dict:
         """
         Load settings from file or environment variables.
 
@@ -174,11 +187,12 @@ class Atomate2Settings(BaseSettings):
         """
         from monty.serialization import loadfn
 
-        config_file_path: str = values.get("CONFIG_FILE", _DEFAULT_CONFIG_FILE_PATH)
+        config_file_path = values.get("CONFIG_FILE", _DEFAULT_CONFIG_FILE_PATH)
+        config_file_path = Path(config_file_path).expanduser()
 
         new_values = {}
-        if Path(config_file_path).expanduser().exists():
-            if Path(config_file_path).stat().st_size == 0:
+        if config_file_path.exists():
+            if config_file_path.stat().st_size == 0:
                 warnings.warn(
                     f"Using atomate2 config file at {config_file_path} but it's empty",
                     stacklevel=2,
