@@ -1,14 +1,20 @@
+from __future__ import annotations
+
 import logging
 from pathlib import Path
-from typing import Dict, Final, Literal, Sequence, Union
+from typing import TYPE_CHECKING, Any, Callable, Final, Literal
 
 import pytest
+from pytest import MonkeyPatch
+
+if TYPE_CHECKING:
+    from collections.abc import Generator, Sequence
 
 logger = logging.getLogger("atomate2")
 
 _VFILES: Final = ("incar", "kpoints", "potcar", "poscar")
-_REF_PATHS: Dict[str, Union[str, Path]] = {}
-_FAKE_RUN_VASP_KWARGS: Dict[str, dict] = {}
+_REF_PATHS: dict[str, str | Path] = {}
+_FAKE_RUN_VASP_KWARGS: dict[str, dict] = {}
 
 
 @pytest.fixture(scope="session")
@@ -22,7 +28,9 @@ def lobster_test_dir(test_dir):
 
 
 @pytest.fixture()
-def mock_vasp(monkeypatch, vasp_test_dir):
+def mock_vasp(
+    monkeypatch: MonkeyPatch, vasp_test_dir: Path
+) -> Generator[Callable[[Any, Any], Any], None, None]:
     """
     This fixture allows one to mock (fake) running VASP.
 
@@ -74,7 +82,14 @@ def mock_vasp(monkeypatch, vasp_test_dir):
         from jobflow import CURRENT_JOB
 
         name = CURRENT_JOB.job.name
-        ref_path = vasp_test_dir / _REF_PATHS[name]
+        try:
+            ref_path = vasp_test_dir / _REF_PATHS[name]
+        except KeyError:
+            raise ValueError(
+                f"no reference directory found for job {name!r}; "
+                f"reference paths received={_REF_PATHS}"
+            ) from None
+
         fake_run_vasp(ref_path, **_FAKE_RUN_VASP_KWARGS.get(name, {}))
 
     get_input_set_orig = VaspInputGenerator.get_input_set
@@ -93,11 +108,8 @@ def mock_vasp(monkeypatch, vasp_test_dir):
     monkeypatch.setattr(VaspInputGenerator, "get_nelect", mock_get_nelect)
 
     def _run(ref_paths, fake_run_vasp_kwargs=None):
-        if fake_run_vasp_kwargs is None:
-            fake_run_vasp_kwargs = {}
-
         _REF_PATHS.update(ref_paths)
-        _FAKE_RUN_VASP_KWARGS.update(fake_run_vasp_kwargs)
+        _FAKE_RUN_VASP_KWARGS.update(fake_run_vasp_kwargs or {})
 
     yield _run
 
@@ -108,7 +120,7 @@ def mock_vasp(monkeypatch, vasp_test_dir):
 
 def fake_run_vasp(
     ref_path: Path,
-    incar_settings: Sequence[str] = (),
+    incar_settings: Sequence[str] = None,
     check_inputs: Sequence[Literal["incar", "kpoints", "poscar", "potcar"]] = _VFILES,
     clear_inputs: bool = True,
 ):
@@ -121,7 +133,8 @@ def fake_run_vasp(
         Path to reference directory with VASP input files in the folder named 'inputs'
         and output files in the folder named 'outputs'.
     incar_settings
-        A list of INCAR settings to check.
+        A list of INCAR settings to check. Defaults to None which checks all settings.
+        Empty list or tuple means no settings will be checked.
     check_inputs
         A list of vasp input files to check. Supported options are "incar", "kpoints",
         "poscar", "potcar", "wavecar".
@@ -160,14 +173,17 @@ def fake_run_vasp(
 def check_incar(ref_path: Path, incar_settings: Sequence[str]):
     from pymatgen.io.vasp import Incar
 
-    user = Incar.from_file("INCAR")
-    ref = Incar.from_file(ref_path / "inputs" / "INCAR")
+    user_incar = Incar.from_file("INCAR")
+    ref_incar_path = ref_path / "inputs" / "INCAR"
+    ref_incar = Incar.from_file(ref_incar_path)
     defaults = {"ISPIN": 1, "ISMEAR": 1, "SIGMA": 0.2}
-    for p in incar_settings:
-        if user.get(p, defaults.get(p)) != ref.get(p, defaults.get(p)):
+    for key in list(user_incar) if incar_settings is None else incar_settings:
+        user_val = user_incar.get(key, defaults.get(key))
+        ref_val = ref_incar.get(key, defaults.get(key))
+        if user_val != ref_val:
             raise ValueError(
-                f"INCAR value of {p} is inconsistent: "
-                f"{user.get(p, defaults.get(p))} != {ref.get(p, defaults.get(p))}"
+                f"\n\nINCAR value of {key} is inconsistent: expected {ref_val}, "
+                f"got {user_val} \nin ref file {ref_incar_path}"
             )
 
 
@@ -188,18 +204,27 @@ def check_kpoints(ref_path: Path):
             "a KPOINTS file"
         )
     if user_kpoints_exists and ref_kpoints_exists:
-        user = Kpoints.from_file("KPOINTS")
-        ref = Kpoints.from_file(ref_path / "inputs" / "KPOINTS")
-        if user.style != ref.style or user.num_kpts != ref.num_kpts:
-            raise ValueError("KPOINTS files are inconsistent")
+        user_kpts = Kpoints.from_file("KPOINTS")
+        ref_kpt_path = ref_path / "inputs" / "KPOINTS"
+        ref_kpts = Kpoints.from_file(ref_kpt_path)
+        if user_kpts.style != ref_kpts.style or user_kpts.num_kpts != ref_kpts.num_kpts:
+            raise ValueError(
+                f"\n\nKPOINTS files are inconsistent: {user_kpts.style} != "
+                f"{ref_kpts.style} or {user_kpts.num_kpts} != {ref_kpts.num_kpts}\nin "
+                f"ref file {ref_kpt_path}"
+            )
     else:
         # check k-spacing
-        user = Incar.from_file("INCAR")
-        ref = Incar.from_file(ref_path / "inputs" / "INCAR")
+        user_incar = Incar.from_file("INCAR")
+        ref_incar_path = ref_path / "inputs" / "INCAR"
+        ref_incar = Incar.from_file(ref_incar_path)
 
-        user_ksp, ref_ksp = user.get("KSPACING"), ref.get("KSPACING")
+        user_ksp, ref_ksp = user_incar.get("KSPACING"), ref_incar.get("KSPACING")
         if user_ksp != ref_ksp:
-            raise ValueError(f"KSPACING is not consistent: {user_ksp} != {ref_ksp}")
+            raise ValueError(
+                f"\n\nKSPACING is inconsistent: expected {ref_ksp}, got {user_ksp} "
+                f"\nin ref file {ref_incar_path}"
+            )
 
 
 def check_poscar(ref_path: Path):
@@ -207,39 +232,47 @@ def check_poscar(ref_path: Path):
     from pymatgen.io.vasp import Poscar
     from pymatgen.util.coord import pbc_diff
 
-    user = Poscar.from_file("POSCAR")
-    ref = Poscar.from_file(ref_path / "inputs" / "POSCAR")
+    user_poscar = Poscar.from_file("POSCAR")
+    ref_poscar = Poscar.from_file(ref_path / "inputs" / "POSCAR")
 
+    user_frac_coords = user_poscar.structure.frac_coords
+    ref_frac_coords = ref_poscar.structure.frac_coords
     if (
-        user.natoms != ref.natoms
-        or user.site_symbols != ref.site_symbols
-        or np.any(
-            np.abs(pbc_diff(user.structure.frac_coords, ref.structure.frac_coords))
-            > 1e-3
-        )
+        user_poscar.natoms != ref_poscar.natoms
+        or user_poscar.site_symbols != ref_poscar.site_symbols
+        or np.any(np.abs(pbc_diff(user_frac_coords, ref_frac_coords)) > 1e-3)
     ):
-        raise ValueError("POSCAR files are inconsistent")
+        ref_poscar_path = ref_path / "inputs" / "POSCAR"
+        user_poscar_path = Path("POSCAR").absolute()
+        raise ValueError(
+            f"POSCAR files are inconsistent\n\n{ref_poscar_path!s}\n{ref_poscar}"
+            f"\n\n{user_poscar_path!s}\n{user_poscar}"
+        )
 
 
 def check_potcar(ref_path: Path):
     from pymatgen.io.vasp import Potcar
 
     if Path(ref_path / "inputs" / "POTCAR").exists():
-        ref = Potcar.from_file(ref_path / "inputs" / "POTCAR").symbols
+        ref_potcar = Potcar.from_file(ref_path / "inputs" / "POTCAR").symbols
     elif Path(ref_path / "inputs" / "POTCAR.spec").exists():
-        ref = Path(ref_path / "inputs" / "POTCAR.spec").read_text().strip().split("\n")
+        ref_potcar = (
+            Path(ref_path / "inputs" / "POTCAR.spec").read_text().strip().split("\n")
+        )
     else:
         raise FileNotFoundError("no reference POTCAR or POTCAR.spec file found")
 
     if Path("POTCAR").exists():
-        user = Potcar.from_file("POTCAR").symbols
+        user_potcar = Potcar.from_file("POTCAR").symbols
     elif Path("POTCAR.spec").exists():
-        user = Path("POTCAR.spec").read_text().strip().split("\n")
+        user_potcar = Path("POTCAR.spec").read_text().strip().split("\n")
     else:
         raise FileNotFoundError("no POTCAR or POTCAR.spec file found")
 
-    if user != ref:
-        raise ValueError(f"POTCAR files are inconsistent: {user} != {ref}")
+    if user_potcar != ref_potcar:
+        raise ValueError(
+            f"POTCAR files are inconsistent: {user_potcar} != {ref_potcar}"
+        )
 
 
 def clear_vasp_inputs():
