@@ -15,8 +15,6 @@ from atomate2.utils.path import strip_hostname
 if TYPE_CHECKING:
     from pathlib import Path
 
-__all__ = ["copy_amset_files"]
-
 
 logger = logging.getLogger(__name__)
 
@@ -26,7 +24,7 @@ def copy_amset_files(
     src_dir: Path | str,
     src_host: str | None = None,
     file_client: FileClient = None,
-):
+) -> None:
     """
     Copy AMSET files to current directory.
 
@@ -80,7 +78,7 @@ def copy_amset_files(
     logger.info("Finished copying inputs")
 
 
-def write_amset_settings(settings_updates: dict, from_prev: bool = False):
+def write_amset_settings(settings_updates: dict, from_prev: bool = False) -> None:
     """
     Write AMSET settings to file.
 
