@@ -7,15 +7,6 @@ from pathlib import Path
 
 from atomate2.utils.file_client import FileClient, auto_fileclient
 
-__all__ = [
-    "copy_files",
-    "delete_files",
-    "rename_files",
-    "gzip_files",
-    "gunzip_files",
-    "get_zfile",
-]
-
 
 @auto_fileclient
 def copy_files(
@@ -28,7 +19,7 @@ def copy_files(
     prefix: str = "",
     allow_missing: bool = False,
     file_client: FileClient | None = None,
-):
+) -> None:
     r"""
     Copy files between source and destination folders.
 
@@ -86,7 +77,7 @@ def delete_files(
     exclude_files: list[str | Path] | None = None,
     allow_missing: bool = False,
     file_client: FileClient | None = None,
-):
+) -> None:
     r"""
     Delete files in a directory.
 
@@ -133,7 +124,7 @@ def rename_files(
     host: str | None = None,
     allow_missing: bool = False,
     file_client: FileClient | None = None,
-):
+) -> None:
     """
     Delete files in a directory.
 
@@ -177,7 +168,7 @@ def gzip_files(
     allow_missing: bool = False,
     force: bool = False,
     file_client: FileClient = None,
-):
+) -> None:
     r"""
     Gzip files in a directory.
 
@@ -230,7 +221,7 @@ def gunzip_files(
     allow_missing: bool = False,
     force: bool = False,
     file_client: FileClient | None = None,
-):
+) -> None:
     r"""
     Gunzip files in a directory.
 
@@ -367,7 +358,7 @@ def get_zfile(
 
 def gzip_output_folder(
     directory: str | Path, setting: bool | str, files_list: list[str]
-):
+) -> None:
     """
     Zip the content of the output folder based on the specific code setting.
 
