@@ -12,8 +12,6 @@ from atomate2.cp2k.schemas.task import TaskDocument
 
 logger = logging.getLogger(__name__)
 
-__all__ = ["Cp2kDrone"]
-
 
 class Cp2kDrone(AbstractDrone):
     """
@@ -25,7 +23,7 @@ class Cp2kDrone(AbstractDrone):
         Additional keyword args passed to :obj:`.TaskDocument.from_directory`.
     """
 
-    def __init__(self, **task_document_kwargs):
+    def __init__(self, **task_document_kwargs) -> None:
         self.task_document_kwargs = task_document_kwargs
 
     def assimilate(self, path: str | Path | None = None) -> TaskDocument:
