@@ -26,16 +26,6 @@ if TYPE_CHECKING:
 
     from atomate2.cp2k.jobs.base import BaseCp2kMaker
 
-__all__ = [
-    "DoubleRelaxMaker",
-    "BandStructureMaker",
-    "RelaxBandStructureMaker",
-    "HybridFlowMaker",
-    "HybridStaticFlowMaker",
-    "HybridRelaxFlowMaker",
-    "HybridCellOptFlowMaker",
-]
-
 
 @dataclass
 class DoubleRelaxMaker(Maker):
@@ -56,7 +46,9 @@ class DoubleRelaxMaker(Maker):
     relax_maker1: Maker = field(default_factory=RelaxMaker)
     relax_maker2: Maker = field(default_factory=RelaxMaker)
 
-    def make(self, structure: Structure, prev_cp2k_dir: str | Path | None = None):
+    def make(
+        self, structure: Structure, prev_cp2k_dir: str | Path | None = None
+    ) -> Flow:
         """
         Create a flow with two chained relaxations.
 
@@ -83,7 +75,7 @@ class DoubleRelaxMaker(Maker):
         return Flow([relax1, relax2], relax2.output, name=self.name)
 
     @classmethod
-    def from_relax_maker(cls, relax_maker: BaseCp2kMaker):
+    def from_relax_maker(cls, relax_maker: BaseCp2kMaker) -> DoubleRelaxMaker:
         """
         Instantiate the DoubleRelaxMaker with two relax makers of the same type.
 
@@ -122,7 +114,9 @@ class BandStructureMaker(Maker):
     static_maker: Maker = field(default_factory=StaticMaker)
     bs_maker: Maker = field(default_factory=NonSCFMaker)
 
-    def make(self, structure: Structure, prev_cp2k_dir: str | Path | None = None):
+    def make(
+        self, structure: Structure, prev_cp2k_dir: str | Path | None = None
+    ) -> Flow:
         """
         Create a band structure flow.
 
@@ -142,7 +136,8 @@ class BandStructureMaker(Maker):
         jobs = [static_job]
 
         outputs = {}
-        if self.bandstructure_type in ("both", "uniform"):
+        bandstructure_type = self.bandstructure_type
+        if bandstructure_type in ("both", "uniform"):
             uniform_job = self.bs_maker.make(
                 static_job.output.structure,
                 prev_cp2k_dir=static_job.output.dir_name,
@@ -156,7 +151,7 @@ class BandStructureMaker(Maker):
             }
             outputs.update(output)
 
-        if self.bandstructure_type in ("both", "line"):
+        if bandstructure_type in ("both", "line"):
             line_job = self.bs_maker.make(
                 static_job.output.structure,
                 prev_cp2k_dir=static_job.output.dir_name,
@@ -170,10 +165,8 @@ class BandStructureMaker(Maker):
             }
             outputs.update(output)
 
-        if self.bandstructure_type not in ("both", "line", "uniform"):
-            raise ValueError(
-                f"Unrecognised bandstructure type {self.bandstructure_type}"
-            )
+        if bandstructure_type not in ("both", "line", "uniform"):
+            raise ValueError(f"Unrecognised {bandstructure_type=}")
 
         return Flow(jobs, outputs, name=self.name)
 
@@ -199,7 +192,9 @@ class RelaxBandStructureMaker(Maker):
     relax_maker: Maker = field(default_factory=DoubleRelaxMaker)
     band_structure_maker: Maker = field(default_factory=BandStructureMaker)
 
-    def make(self, structure: Structure, prev_cp2k_dir: str | Path | None = None):
+    def make(
+        self, structure: Structure, prev_cp2k_dir: str | Path | None = None
+    ) -> Flow:
         """
         Run a relaxation and then calculate the uniform and line mode band structures.
 
@@ -248,7 +243,7 @@ class HybridFlowMaker(Maker):
     pbe_maker: Maker = field(default=lambda: StaticMaker(store_output_data=False))
     hybrid_maker: Maker = field(default_factory=HybridStaticMaker)
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         """
         Post init updates.
 
@@ -259,7 +254,9 @@ class HybridFlowMaker(Maker):
         the PBE density matrix, which creates huge speed-ups. Rarely causes
         problems so it is done as a default here.
         """
-        updates = {"activate_hybrid": {"hybrid_functional": self.hybrid_functional}}
+        updates: dict[str, dict[str, str | bool]] = {
+            "activate_hybrid": {"hybrid_functional": self.hybrid_functional}
+        }
         if self.initialize_with_pbe:
             updates["activate_hybrid"].update(
                 {"screen_on_initial_p": True, "screen_p_forces": True}
