@@ -5,7 +5,7 @@ from __future__ import annotations
 import logging
 import re
 from pathlib import Path
-from typing import TYPE_CHECKING, Sequence
+from typing import TYPE_CHECKING
 
 from atomate2 import SETTINGS
 from atomate2.common.files import copy_files, get_zfile, gunzip_files, rename_files
@@ -13,11 +13,11 @@ from atomate2.utils.file_client import FileClient, auto_fileclient
 from atomate2.utils.path import strip_hostname
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     from pymatgen.core import Structure
 
     from atomate2.vasp.sets.base import VaspInputGenerator
-
-__all__ = ["copy_vasp_outputs", "get_largest_relax_extension"]
 
 
 logger = logging.getLogger(__name__)
@@ -31,7 +31,7 @@ def copy_vasp_outputs(
     contcar_to_poscar: bool = True,
     force_overwrite: bool | str = False,
     file_client: FileClient | None = None,
-):
+) -> None:
     """
     Copy VASP output files to the current directory.
 
@@ -84,7 +84,7 @@ def copy_vasp_outputs(
 
     # check at least one type of POTCAR file is included
     if len([f for f in optional_files if "POTCAR" in f.name]) == 0:
-        raise FileNotFoundError("Could not find POTCAR file to copy.")
+        raise FileNotFoundError(f"Could not find a POTCAR file in {src_dir!r} to copy")
 
     copy_files(
         src_dir,
@@ -161,7 +161,7 @@ def write_vasp_input_set(
     potcar_spec: bool = False,
     clean_prev: bool = True,
     **kwargs,
-):
+) -> None:
     """
     Write VASP input set.
 
