@@ -42,9 +42,6 @@ if TYPE_CHECKING:
     from atomate2.vasp.jobs.base import BaseVaspMaker
 
 
-__all__ = ["ElectronPhononMaker"]
-
-
 @dataclass
 class ElectronPhononMaker(Maker):
     """
@@ -126,7 +123,9 @@ class ElectronPhononMaker(Maker):
         )
     )
 
-    def make(self, structure: Structure, prev_vasp_dir: str | Path | None = None):
+    def make(
+        self, structure: Structure, prev_vasp_dir: str | Path | None = None
+    ) -> Flow:
         """
         Create a electron-phonon coupling workflow.
 
