@@ -25,7 +25,6 @@ if TYPE_CHECKING:
 
     from atomate2.vasp.jobs.base import BaseVaspMaker
 
-__all__ = ["VaspLobsterMaker"]
 
 LOBSTER_UNIFORM_MAKER = UniformBandStructureMaker(
     name="uniform lobster structure",
@@ -49,7 +48,8 @@ LOBSTER_UNIFORM_MAKER = UniformBandStructureMaker(
                 "LWAVE": True,
                 "ISYM": 0,
             },
-        )
+        ),
+        task_document_kwargs={"parse_dos": False, "parse_bandstructure": False},
     ),
 )
 
@@ -62,9 +62,8 @@ class VaspLobsterMaker(Maker):
     The calculations performed are:
 
     1. Optional optimization.
-    2. Optional static computation with symmetry to preconverge the wavefunction.
-    3. Static calculation with ISYM=0.
-    4. Several Lobster computations testing several basis sets are performed.
+    2. Static calculation with ISYM=0.
+    3. Several Lobster computations testing several basis sets are performed.
 
     .. Note::
 
@@ -77,9 +76,6 @@ class VaspLobsterMaker(Maker):
     relax_maker : .BaseVaspMaker or None
         A maker to perform a relaxation on the bulk. Set to ``None`` to skip the
         bulk relaxation.
-    preconverge_static_maker : .BaseVaspMaker or None
-        A maker to perform a preconvergence run before the wavefunction computation
-        without symmetry
     lobster_static_maker : .BaseVaspMaker
         A maker to perform the computation of the wavefunction before the static run.
         Cannot be skipped. It can be LOBSTERUNIFORM or LobsterStaticMaker()
@@ -109,7 +105,7 @@ class VaspLobsterMaker(Maker):
         self,
         structure: Structure,
         prev_vasp_dir: str | Path | None = None,
-    ):
+    ) -> Flow:
         """
         Make flow to calculate bonding properties.
 
