@@ -11,6 +11,54 @@ from pymatgen.io.vasp import Kpoints
 from atomate2.vasp.jobs.base import BaseVaspMaker
 
 
+def update_vasp_input_generators(
+    flow: Job | Flow | Maker,
+    dict_mod_updates: dict[str, Any],
+    name_filter: str | None = None,
+    class_filter: type[Maker] | None = BaseVaspMaker,
+) -> Job | Flow | Maker:
+    """
+    Update any VaspInputGenerators or Makers in the flow.
+
+    Note, this returns a copy of the original Job/Flow/Maker. I.e., the update does not
+    happen in place.
+
+    Parameters
+    ----------
+    flow : .Job or .Flow or .Maker
+        A job, flow or Maker.
+    dict_mod_updates : dict
+        The updates to apply. Existing keys will not be modified unless explicitly
+        specified in ``dict_mod_updates``.
+    name_filter : str or None
+        A filter for the name of the jobs.
+    class_filter : Maker or None
+        A filter for the VaspMaker class used to generate the flows. Note the class
+        filter will match any subclasses.
+
+    Returns
+    -------
+    Job or Flow or Maker
+        A copy of the input flow/job/maker modified to use the updated incar settings.
+    """
+    updated_flow = deepcopy(flow)
+    if isinstance(updated_flow, Maker):
+        updated_flow = updated_flow.update_kwargs(
+            {"_set": dict_mod_updates},
+            name_filter=name_filter,
+            class_filter=class_filter,
+            dict_mod=True,
+        )
+    else:
+        updated_flow.update_maker_kwargs(
+            {"_set": dict_mod_updates},
+            name_filter=name_filter,
+            class_filter=class_filter,
+            dict_mod=True,
+        )
+    return updated_flow
+
+
 def update_user_incar_settings(
     flow: Job | Flow | Maker,
     incar_updates: dict[str, Any],
@@ -44,26 +92,15 @@ def update_user_incar_settings(
     Job or Flow or Maker
         A copy of the input flow/job/maker modified to use the updated incar settings.
     """
-    dict_mod_updates = {
-        f"input_set_generator->user_incar_settings->{k}": v
-        for k, v in incar_updates.items()
-    }
-    updated_flow = deepcopy(flow)
-    if isinstance(updated_flow, Maker):
-        updated_flow = updated_flow.update_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    else:
-        updated_flow.update_maker_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    return updated_flow
+    return update_vasp_input_generators(
+        flow=flow,
+        dict_mod_updates={
+            f"input_set_generator->user_incar_settings->{k}": v
+            for k, v in incar_updates.items()
+        },
+        name_filter=name_filter,
+        class_filter=class_filter,
+    )
 
 
 def update_user_potcar_settings(
@@ -99,26 +136,15 @@ def update_user_potcar_settings(
     Job or Flow or Maker
         A copy of the input flow/job/maker modified to use the updated potcar settings.
     """
-    dict_mod_updates = {
-        f"input_set_generator->user_potcar_settings->{k}": v
-        for k, v in potcar_updates.items()
-    }
-    updated_flow = deepcopy(flow)
-    if isinstance(updated_flow, Maker):
-        updated_flow = updated_flow.update_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    else:
-        updated_flow.update_maker_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    return updated_flow
+    return update_vasp_input_generators(
+        flow=flow,
+        dict_mod_updates={
+            f"input_set_generator->user_potcar_settings->{k}": v
+            for k, v in potcar_updates.items()
+        },
+        name_filter=name_filter,
+        class_filter=class_filter,
+    )
 
 
 def update_user_potcar_functional(
@@ -153,25 +179,14 @@ def update_user_potcar_functional(
     Job or Flow or Maker
         A copy of the input flow/job/maker modified to use the updated potcar settings.
     """
-    dict_mod_updates = {
-        "input_set_generator->user_potcar_functional": potcar_functional
-    }
-    updated_flow = deepcopy(flow)
-    if isinstance(updated_flow, Maker):
-        updated_flow = updated_flow.update_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    else:
-        updated_flow.update_maker_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    return updated_flow
+    return update_vasp_input_generators(
+        flow=flow,
+        dict_mod_updates={
+            "input_set_generator->user_potcar_functional": potcar_functional
+        },
+        name_filter=name_filter,
+        class_filter=class_filter,
+    )
 
 
 def update_user_kpoints_settings(
@@ -217,23 +232,12 @@ def update_user_kpoints_settings(
             f"input_set_generator->user_kpoints_settings->{k}": v
             for k, v in kpoints_updates.items()
         }
-
-    updated_flow = deepcopy(flow)
-    if isinstance(updated_flow, Maker):
-        updated_flow = updated_flow.update_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    else:
-        updated_flow.update_maker_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    return updated_flow
+    return update_vasp_input_generators(
+        flow=flow,
+        dict_mod_updates=dict_mod_updates,
+        name_filter=name_filter,
+        class_filter=class_filter,
+    )
 
 
 def use_auto_ispin(
@@ -267,21 +271,9 @@ def use_auto_ispin(
     Job or Flow or Maker
         A copy of the input flow/job/maker but with auto_ispin set.
     """
-    dict_mod_updates = {"input_set_generator->auto_ispin": value}
-
-    updated_flow = deepcopy(flow)
-    if isinstance(updated_flow, Maker):
-        updated_flow = updated_flow.update_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    else:
-        updated_flow.update_maker_kwargs(
-            {"_set": dict_mod_updates},
-            name_filter=name_filter,
-            class_filter=class_filter,
-            dict_mod=True,
-        )
-    return updated_flow
+    return update_vasp_input_generators(
+        flow=flow,
+        dict_mod_updates={"input_set_generator->auto_ispin": value},
+        name_filter=name_filter,
+        class_filter=class_filter,
+    )
