@@ -1,3 +1,6 @@
+import pytest
+
+
 def test_elph_renormalisation(mock_vasp, clean_dir, si_structure):
     from jobflow import run_locally
 
@@ -65,8 +68,5 @@ def test_elph_renormalisation(mock_vasp, clean_dir, si_structure):
     # validation on the outputs
     renorm_output = responses[flow.output.uuid][1].output
     assert isinstance(renorm_output, ElectronPhononRenormalisationDoc)
-    assert set(renorm_output.delta_band_gaps) == {
-        -0.488900000000001,
-        -0.48850000000000104,
-    }
+    assert renorm_output.delta_band_gaps == pytest.approx([-0.4889, -0.4885], rel=1e-3)
     assert renorm_output.chemsys == "Si"
